Extracted function into separate file, minor string updates
Took the dateConverter function and placed it into its own file for re-usability. Also made some updates to string concatenation in user_profile
